requirements.txt
asgiref==3.7.2
dj-database-url==0.5.0
Django==4.2.10
django-autoslug==1.9.9
gunicorn==20.1.0
psycopg2==2.9.9
sqlparse==0.4.4
whitenoise==6.5.0
